requirements.txt
scrapfly-sdk
beautifulsoup4
# sqlite3
loguru
discord.py[voice]
python-dotenv